requirements.txt
aiocron==1.8
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.10b0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
croniter==1.0.15
discord==1.7.3
discord.py==1.7.3
idna==3.3
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.2
six==1.16.0
tomli==1.2.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
yarl==1.7.2
